Split entity utilities to a new module
I don't want to expose of all of misc-utils as public API in Genie
(it was before but it was a mistake), so extract the few methods
that almond-cloud will need to process datasets into a separate
file.
