+++ lib/index.js
@@ -41,6 +41,7 @@ const ParserClient = require('./prediction/parserclient');
 
 const I18n = require('./i18n');
 const parallelize = require('./utils/parallelize');
+const EntityUtils = require('./utils/entity-utils');
 
 const SpeechHandler = require('./speech/speech_handler');
 const AssistantEngine = require('./engine');
@@ -88,6 +89,7 @@ module.exports = {
 
     // semi-unstable API
     parallelize,
+    EntityUtils,
 
     // dialogue agent
     SpeechHandler,
